policies added

+++ modules/Fateme/Category/Http/Controllers/CategoryController.php
@@ -18,6 +18,7 @@ class CategoryController  extends Controller
     public function index()
     {
 //        dd(auth()->user()->permissions);
+        $this->authorize('manage', Category::class);
         $categories = $this->repo->all();
         return view('Categories::index', compact('categories'));
 
@@ -25,12 +26,14 @@ class CategoryController  extends Controller
 
     public function store(CategoryRequest $request)
     {
+        $this->authorize('manage', Category::class);
          $this->repo->store($request);
             return back();
     }
 
     public function edit($categoryId)
     {
+        $this->authorize('manage', Category::class);
         $category = $this->repo->findById($categoryId);
         $categories = $this->repo->allExceptById($categoryId);
         return view('Categories::edit', compact('category', 'categories'));
@@ -38,12 +41,14 @@ class CategoryController  extends Controller
 
     public function update($categoryId, CategoryRequest $request)
     {
+        $this->authorize('manage', Category::class);
         $this->repo->update($categoryId, $request);
         return back();
     }
 
     public function destroy($categoryId)
     {
+        $this->authorize('manage', Category::class);
         $this->repo->delete($categoryId);
        return AjaxResponses::SuccessResponse();
     }
